requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
beautifulsoup4>=4.11.0
umap-learn>=0.5.3
voyageai>=0.2.0
pathlib2>=2.3.7; python_version<"3.4"
pandas>=1.3.0
scikit-learn>=0.24.0
networkx>=2.6.0
playwright>=1.40.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
ftfy
regex
tqdm 